Rename GraphObject to GraphNode
