+++ src/Facebook/GraphNodes/GraphApplication.php
@@ -29,7 +29,7 @@ namespace Facebook\GraphNodes;
  * @package Facebook
  */
 
-class GraphApplication extends GraphObject
+class GraphApplication extends GraphNode
 {
     /**
      * Returns the ID for the application.
